requirements.txt
zipp==3.17.0
werkzeug==3.0.1
urllib3==2.1.0
tzdata==2023.3
typing-extensions==4.8.0
six==1.16.0
setuptools==56.0.0
requests==2.31.0
pytz==2023.3.post1
python-dotenv==1.0.0
python-dateutil==2.8.2
pyjwt==2.8.0
pydantic==1.10.13
pycparser==2.21
pip==21.1.1
pandas==2.1.3
openpyxl==3.1.2
numpy==1.26.2
markupsafe==2.1.3
jinja2==3.1.2
itsdangerous==2.1.2
importlib-metadata==6.8.0
idna==3.4
flask==3.0.0
fhir.resources==6.0.0
et-xmlfile==1.1.0
email-validator==2.1.0.post1
dnspython==2.4.2
curlify==2.2.1
cryptography==41.0.5
click==8.1.7
charset-normalizer==3.3.2
cffi==1.16.0
certifi==2023.11.17
blinker==1.7.0